chore: fix defmt

+++ src/fat/volume.rs
@@ -595,7 +595,7 @@ impl FatVolume {
                 #[cfg(feature = "defmt-log")]
                 debug!(
                     "LFN Contents {=u8} {=u8} {=u8:02x} {=[?; 13]:#04x}",
-                    start, sequence, csum, buffer
+                    start as u8, sequence, csum, buffer
                 );
                 match (start, sequence, self) {
                     (true, 0x01, _) => {
